[ADQL,TAP] Add the MOC and MOC_AGG functions in the ADQL grammar.
There is also a new type of type geometry: MOC. This should allow logical
operations between a MOC and a POINT: CONTAINS and INTERSECTS.

The function AREA(...) is also supported, but partially: it will be called
only if the given geometry is explicitely a MOC (i.e. not a column). To fix
this problem, the implementation in pg_moc should be changed:
area(MOC) should return an area in square radians instead of square degrees.

Since PgSphere is required to use pg_moc (for the moment), the PgSphere
translator has been extented to make: PgMocTranslator. This translator can
be referenced in tap.properties with the special value "pgmoc".