[UWS] Fix warnings while restoring a backup.
The attribute `version` in the node `job` is never restored ; it is just
informative. Thus, it is now silently ignored.

The attribute `length` of an upload was badly spelled.

The attribute `mime-type` of a result was backuped as `mime-type` but restored
as `mime`. Hence the absence of this piece of information when restoring a job.