[ADQL] In the FROM, set a default alias for each table reference (i.e. not a subquery).
The idea is to avoid ambiguous error messages coming from the database when
raising an error on a table or column name having different DB and ADQL names.

This commit sets by default an alias on each table reference. This default
alias is the ADQL table name as it is used in the ADQL query.

_This commit fixes the GitHub issue #108 ._