[ADQL] Abort previous commit (09fac540f7ab7d8c92150fd7b0d52dfe21a46eb2).
The previous commit broke some JUnit tests. It was indeed not any more possible
to prefix a column with the fully qualified table name
(e.g. `"aSchema"."aTable"`) if the table was automatically aliased. If an alias
is set on a table, this table should be referenced only by this alias. The
previous commit was not able to replace fully qualified table names with their
corresponding alias....and doing that would probably introduce other unexpected
parsing/checking behaviors.

So, it is better to keep this part of the parsing as simple as now. Just for few
DB error messages showing real database names (and only if they are different
from the ADQL ones), it does not worth _yet_ making the parsing more complex.

_This commit ends the issue #108 ._