[UWS] Fix the parsing of the HTTP header `Accept`.
In case a MIME-type parameter was not `q` set to a floating point value
(e.g. correct is `q=0.8` ; incorrect is `q=abc`), the library was throwing an
ugly NumberFormatException. This exception is now caught (and ignored) if it
occurs.

The same exception was also thrown for any other parameter whose value is not a
floating point. Since only the quality flag (i.e. `q`) is used in UWS-Lib,
parameters are now only parsed if it is `q` ; all others are now ignored.